+++ sphinxcontrib/osint/scripts/index.py
@@ -15,7 +15,7 @@ import sys
 import click
 import pycountry
 
-from ..xapian import XapianIndexer, XapianHTMLIndexer
+from ..xapian import XapianIndexer, context_data
 from . import parser_makefile, cli, get_app, load_quest
 
 @cli.command()
@@ -43,32 +43,16 @@ def build(common):
 @cli.command()
 @click.option('--fuzzy/--no-fuzzy', default=True, help="Use fuzzy search")
 @click.option('--threshold', default=50, help="Similarity threshold for fuzzy search (0-100)")
-@click.option('--limit', default=10, help="Maximum number of results")
+@click.option('--limit', default=10, help="Results per page")
+@click.option('--offset', default=0, help="Offset for results")
 @click.option('--home', default='http://127.0.0.1:8000/', help="Maximum number of results")
 @click.option('--types', default=None, help="Types of data to search")
 @click.option('--cats', default=None, help="Cats of data to search")
 @click.option('--countries', default=None, help="Countries of data to search")
 @click.argument('query', default=None)
 @click.pass_obj
-def search(common, fuzzy, threshold, limit, home, types, cats, countries, query):
+def search(common, fuzzy, threshold, offset, limit, home, types, cats, countries, query):
     """Search"""
-
-    def print_data(searches, data, distance=60):
-        ret = ''
-        for search in searches.split(' '):
-            idx = data.lower().find(search.lower())
-            if idx != -1:
-                dist_min = idx - distance
-                if dist_min < 0:
-                    dist_min = 0
-                dist_max = idx + distance
-                if dist_max > len(data):
-                    dist_max = len(data)
-                if ret != '':
-                    ret += '...'
-                ret += data[dist_min:dist_max]
-        return ret
-
     sourcedir, builddir = parser_makefile(common.docdir)
     app = get_app(sourcedir=sourcedir, builddir=builddir)
 
@@ -85,20 +69,22 @@ def search(common, fuzzy, threshold, limit, home, types, cats, countries, query)
 
     results = indexer.search(query,
         use_fuzzy=fuzzy, fuzzy_threshold=threshold,
-        limit=limit, cats=cats, types=types, countries=countries)
+        limit=limit, offset=offset,
+        cats=cats, types=types, countries=countries)
 
-    print(f"\n=== Résults for: '{query}' ===")
-    print(f"Found {len(results)}\n")
+    print(f"\n=== Results for: '{results['query']}' ===")
+    print(f"Found : fuzzy:{len(results['results'])} / xapian:{results['total']}\n")
 
-    for result in results:
+    for result in results['results']:
         print(f"[{result['rank']}] {result['title']}")
-        print(f"   URL: {home}{result['filepath']}")
+        print(f"   Link: {home}{result['filepath']}")
+        print(f"   URL : {result['url']}")
         print(f"   Score: {result['score']}%", end='')
         if 'fuzzy_score' in result:
             print(f" | Fuzzy: {result['fuzzy_score']:.1f} | Combiné: {result['combined_score']:.1f}", end='')
         print("")
         print(f"   Type: {result['type']} | Cats: {result['cats']} | Country: {result['country']}")
-        print(f"   Data: ...{print_data(query, result['data'])}...")
+        print(f"   Data: ...{context_data(query, result['data'])}...")
         print("")